+++ .eslintrc.json
@@ -15,6 +15,7 @@
     "import/prefer-default-export": "off",
     "no-alert": "off",
     "no-plusplus": "off",
-    "consistent-return": "off"
+    "consistent-return": "off",
+    "no-restricted-exports": "off"
   }
 }
