+++ src/js/components/ProductManageMenu/index.js
@@ -1,0 +1,1 @@
+export { default } from './ProductManageMenu.js';
