+++ src/js/components/VendingMachineManageMenu/index.js
@@ -1,0 +1,1 @@
+export { default } from './VendingMachineManageMenu.js';
